requirements.txt
pyarrow
pydantic
datasets
pandas
python-dotenv
GitPython